% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fledgingr.R
\name{tokenize}
\alias{tokenize}
\title{Tokenize and Analyze Japanese Texts}
\usage{
tokenize(x, mode = c("C", "B", "A"))
}
\arguments{
\item{x}{Text to tokenize.}

\item{mode}{Unit to split text.
\itemize{
\item \code{"A"}: short
\item \code{"B"}: middle
\item \code{"C"}: long
}}
}
\description{
Tokenize and Analyze Japanese Texts
}
